ecs-codepipeline/requirements.txt
Flask==3.0.2
SQLAlchemy==2.0.29
PyMySQL==1.1.0
python-dotenv==1.0.1
gunicorn==21.2.0
